requirements.txt
python-dotenv==1.0.0
openai==1.3.0
fpdf2==2.7.6
